The elapsed time is now saved and printed correctly in highscores and level end The highscores are now floats rather than ushorts, representing time elapsed in seconds

+++ Assets/Scripts/GlobalControl.cs
@@ -20,7 +20,7 @@ public class GlobalControl : MonoBehaviour
 		/*
 			* the time the played survived in the previous level (for endgame scene)
 			*/
-		public ushort timeSurvived = 0;
+		public float timeSurvived = 0;
 
 		/*
 			* the level that was played (to be used in retry function)
